requirements.txt
pytest==8.3.2
httpx==0.27.0
pydantic==2.8.2
allure-pytest==2.13.5
